requirements.txt
api-display-purposes==0.0.3
APScheduler==3.6.3
asgiref==3.4.1
async-generator==1.10
async-lru==1.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
clarifai==2.6.2
colorama==0.4.4
configparser==3.8.1
cryptography==36.0.1
distlib==0.3.4
Django==4.0
EasyProcess==0.3
emoji==1.6.1
filelock==3.4.2
future==0.18.2
googleapis-common-protos==1.54.0
grpcio==1.43.0
h11==0.12.0
idna==2.10
instapy==0.6.15
jsonschema==2.6.0
MeaningCloud-python==2.0.0
outcome==1.1.0
platformdirs==2.4.1
plyer==2.0.0
protobuf==3.19.1
pyaes==1.6.1
pycparser==2.21
pygame==2.1.2
pymongo==4.0.1
pyOpenSSL==21.0.0
Pyrogram==1.2.20
PySocks==1.7.1
pyTelegramBotAPI==4.2.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
selenium==4.1.0
semantic-version==2.8.5
setuptools-rust==1.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.13.0
webdriverdownloader==1.1.0.3
wsproto==1.0.0
